Add the ability to pause the world scene to show a menu
In this commit, the use case of pausing the world scene is to go back to
the main menu (the only menu that exists at this time). On start-up, the
main menu has no background. With this change, the user will now see a
frozen still of the world scene when the Escape key is pressed during
the world scene.